requirements.txt
langchain
langchain-community
openai
faiss-cpu
pypdf
gradio
python-dotenv
